optimuskg/pipelines/silver/nodes/opentargets/clingen.py
import pandas as pd
import polars as pl
from kedro.pipeline import node

from .utils import construct_edges


def run(  # noqa: PLR0913
    clingen: pd.DataFrame,
    hp_terms: pl.DataFrame,
    diseases: pl.DataFrame,
    targets: pl.DataFrame,
    drug_mappings: pl.DataFrame,
    disease_phenotype_ids: pl.DataFrame,
    score_threshold: float,
) -> pl.DataFrame:
    df = pl.from_pandas(clingen)
    df = (
        df.select(
            [
                "id",
                "targetId",
                "diseaseId",
                "targetFromSourceId",
                "diseaseFromSourceId",
                "score",
            ]
        )
        .filter(
            pl.col("targetId").is_in(targets["id"])
            & pl.col("diseaseId").is_in(disease_phenotype_ids["id"])
            & (pl.col("score") > score_threshold)
        )
        .pipe(
            construct_edges,
            targets_df=pl.DataFrame(targets),
            phenotypes_df=pl.DataFrame(hp_terms),
            diseases_df=diseases,
            drug_mappings_df=drug_mappings,
            type_x="gene",
            type_y="disease",
            relation_label="disease_protein",
            display_relation_label="associated with",
        )
    )
    df = df.sort(by=sorted(df.columns))
    return df


clingen_node = node(
    run,
    inputs={
        "clingen": "bronze.opentargets.evidence.clingen",
        "hp_terms": "bronze.ontology.hp_terms",
        "diseases": "bronze.opentargets.diseases",
        "targets": "bronze.opentargets.targets",
        "disease_phenotype_ids": "bronze.opentargets.disease_phenotype_ids",
        "drug_mappings": "bronze.opentargets.drug_mappings",
        "score_threshold": "params:score_threshold",
    },
    outputs="opentargets.evidence.clingen",
    name="clingen",
    tags=["silver"],
)
