pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
requires-python = ">=3.12"
name = "optimuskg"
readme = "README.md"
version = "0.31.0"
dependencies = [
  "ipython>=8.10",
  "jupyterlab>=3.0",
  "notebook",
  "kedro[jupyter]~=0.19.10",
  "kedro-datasets[pandas-csvdataset, pandas-jsondataset, polars-csvdataset]>=3.0",
  "typedframe>=0.11.0",
  "polars>=1.19.0",
  "biocypher>=0.6.1",
  "requests>=2.32.3",
  "aiohttp>=3.11.11",
  "goatools>=1.4.12",
  "lxml>=5.3.0",
  "pandas-stubs>=2.2.3.241126",
  "types-openpyxl>=3.1.5.20241225",
  "lxml-stubs>=0.5.1",
  "psycopg2-binary>=2.9.10",
  "pyarrow>=19.0.0",
  "sqlalchemy>=2.0.37",
  "types-requests>=2.32.0.20250328",
  "kedro-viz>=11.0.1",
  "owlready2>=0.47",
]

[dependency-groups]
tests = [
  "pytest-cov~=3.0",
  "pytest-mock>=1.7.1, <2.0",
  "pytest~=7.2",
]
dev = [
  "ruff~=0.1.8",
  "pre-commit>=4.0.1",
  "mypy>=1.14.1",
  "bandit[toml]>=1.8.2",
  "interrogate>=1.7.0",
  "pydantic>=2.10.6",
  "typer>=0.15.1",
  "types-cachetools>=5.5.0.20240820",
  "scienceplots>=2.1.1",
]

[project.scripts]
optimuskg = "optimuskg.__main__:main"
cli = "cli.__main__:app"

[tool.hatch.build.targets.wheel]
packages = ["optimuskg"]

[tool.hatch.envs.default]
path = ".venv"

[tool.uv]
required-version = ">=0.5.18"

[tool.kedro]
package_name = "optimuskg"
project_name = "OptimusKG"
kedro_init_version = "0.19.10"
source_dir = "optimuskg"

[tool.pytest.ini_options]
addopts = """
--cov-report term-missing \
--cov optimuskg -ra"""

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.typos.default.extend-words]
# Don't correct the OpenTargets Abbreviation
ot = "ot"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff]
line-length = 88
show-fixes = true
lint.select = [
  "F",    # Pyflakes
  "W",    # pycodestyle
  "E",    # pycodestyle
  "I",    # isort
  "UP",   # pyupgrade
  "PL",   # Pylint
  "T201", # Print Statement
]
lint.ignore = ["E501"] # Ruff format takes care of line-too-long

[tool.kedro_telemetry]
project_id = "dc32d474bbc84fc49bf67c0f87edf02e"

[tool.bandit]
skips = ["B101"]

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = false
fail-under = 10
verbose = 2
quiet = false
color = true

[tool.mypy]
warn_unused_configs = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true


[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
    "fsspec.*",
    "goatools.*",
    "biocypher.*",
    "typedframe.*",
    "pytest.*",
    "pydantic.*",
    "kedro_datasets.partitions.partitioned_dataset.*",
]

[[tool.mypy.overrides]]
module = ["optimuskg.datasets.*"]
disable_error_code = ["attr-defined"]

[tool.typos.files]
extend-exclude = ["tests/stub_data/*", "Makefile", "notebooks/*"]
